fix: refactored the PublicPageModal component and used the Avatar component in ProfileData
