+++ src/shared/ui/carousel/carousel.module.scss
@@ -1,4 +1,4 @@
-.modalSlider {
+.carousel {
   width: 492px;
   position: relative;
   margin: 0 auto;
