+++ src/shared/ui/modalComments/commentsBlock/commentContainer/comment/Comment.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import s from './comment.module.scss'
 type CommentProps = {
   text?: string
