+++ src/shared/ui/modalComments/commentsBlock/index.ts
