+++ src/shared/ui/modalComments/commentsFooter/index.ts
