+++ src/shared/ui/modalComments/datePost/DatePost.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import s from './datePost.module.scss'
 
 type DatePostProps = {
