+++ src/shared/ui/modalComments/profileData/ProfileData.tsx
@@ -1,4 +1,8 @@
+'use client'
+
 import React from 'react'
+
+import { Avatar, AvatarFallback, AvatarImage } from '@/shared/ui/avatar/Avatar'
 
 import s from './profileData.module.scss'
 
@@ -10,7 +14,11 @@ type ProfileDataProps = {
 export const ProfileData: React.FC<ProfileDataProps> = ({ className, imageUrl, profileUrl }) => {
   return (
     <span className={className || s.profileData}>
-      <img alt={'ProfileImg'} className={s.img} src={imageUrl || ''} />
+      {/* <img alt={'ProfileImg'} className={s.img} src={imageUrl || ''} /> */}
+      <Avatar>
+        <AvatarImage alt={'ProfileImg'} className={s.img} src={imageUrl || ''} />
+        <AvatarFallback>A1</AvatarFallback>
+      </Avatar>
       <div className={s.urlProfile}>{profileUrl}</div>
     </span>
   )
