+++ src/shared/ui/modalComments/profileData/index.ts
