--- src/shared/ui/publicPageModal/modalSlider/index.ts
@@ -1,1 +1,0 @@
-export { ModalSlider } from './ModalSlider'
