+++ src/views/publicPageModal/DataArray.tsx
