+++ src/views/publicPageModal/ui/publicPageModal.module.scss
